# swagger.yaml
basePath: /
consumes:
- application/json
...
info:
  description: Documentation for Product API
  title: Product API.
  version: 1.0.0
paths:
...
  /products{id}:
    put:
      description: Updates a product
      operationId: updateProduct
      responses:
        "201":
          $ref: '#/responses/noContent'
        "422":
          $ref: '#/responses/errorValidation'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
produces:
- application/json
responses:
  noContent:
    description: Returned when deleting a product
...
schemes:
- http
- https
swagger: "2.0"